+++ src/components/admin/AdminLayout.jsx
@@ -1,13 +1,12 @@
 import { Outlet } from "react-router-dom";
 import AdminHeader from "./AdminHeader";
 import BottomNavBar from "./BottomNavBar";
-import AdminMain from "./AdminMain";
+
 function AdminLayout() {
   return (
     <div className="w-full overflow-hidden">
       <AdminHeader />
       <Outlet />
-      <AdminMain />
       <BottomNavBar />
     </div>
   );
